# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "GuildPass Access API",
    "description": "MVP API for wallet membership and access checks",
    "version": "0.1.0"
  },
...
  "paths": {
...
    "/v1/wallets/{primaryWallet}/challenges": {
      "post": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "primaryWallet",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/wallets/{primaryWallet}/link": {
      "post": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "primaryWallet",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/wallets/{primaryWallet}/linked": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "primaryWallet",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/memberships/{wallet}/appeals": {
      "post": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "wallet",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/appeals/{appealId}/transition": {
      "post": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "appealId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/communities/{communityId}/memberships/{wallet}": {
      "get": {
        "summary": "Get membership status summary for a wallet in a community",
        "tags": [
          "Memberships"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "communityId",
            "required": true,
            "description": "Community identifier"
          },
          {
            "schema": {
              "type": "string",
              "pattern": "^0x[0-9a-fA-F]{40}$"
            },
            "in": "path",
            "name": "wallet",
            "required": true,
            "description": "EVM-compatible wallet address (checksummed or lowercase)"
          }
        ],
        "responses": {
          "200": {
            "description": "Membership summary for the wallet",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Membership summary for the wallet",
                  "type": "object",
                  "properties": {
                    "wallet": {
                      "type": "string",
                      "pattern": "^0x[0-9a-fA-F]{40}$",
                      "description": "EVM-compatible wallet address (checksummed or lowercase)"
                    },
                    "communities": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "communityId",
                          "state"
                        ],
                        "properties": {
                          "communityId": {
                            "type": "string"
                          },
                          "state": {
                            "type": "string",
                            "enum": [
                              "invited",
                              "active",
                              "expired",
                              "suspended"
                            ]
                          },
                          "expiresAt": {
                            "type": "string",
                            "format": "date-time",
                            "nullable": true
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Wallet not found",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Wallet not found",
                  "type": "object",
                  "required": [
                    "error",
                    "code",
                    "message",
                    "statusCode"
                  ],
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Machine-readable error identifier"
                    },
                    "code": {
                      "type": "string",
                      "description": "HTTP status phrase / error code"
                    },
                    "message": {
                      "type": "string",
                      "description": "Human-readable description"
                    },
                    "statusCode": {
                      "type": "integer",
                      "description": "HTTP status code"
                    },
                    "details": {
                      "description": "Optional detail payload",
                      "oneOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "object"
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Internal server error",
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/communities/{communityId}/overrides/{wallet}/{resource}": {
      "delete": {
        "summary": "Revoke an access override for a wallet/resource pair",
        "tags": [
          "Overrides"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "communityId",
            "required": true,
            "description": "Community identifier"
          },
          {
            "schema": {
              "type": "string",
              "pattern": "^0x[0-9a-fA-F]{40}$"
            },
            "in": "path",
            "name": "wallet",
            "required": true,
            "description": "EVM-compatible wallet address (checksummed or lowercase)"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "resource",
            "required": true,
            "description": "Resource identifier"
          }
        ],
        "responses": {
          "200": {
            "description": "Override revoked",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Override revoked",
                  "type": "object",
                  "required": [
                    "communityId",
                    "wallet",
                    "resource",
                    "effect",
                    "created",
                    "removed"
                  ],
                  "properties": {
                    "communityId": {
                      "type": "string"
                    },
                    "wallet": {
                      "type": "string",
                      "pattern": "^0x[0-9a-fA-F]{40}$",
                      "description": "EVM-compatible wallet address (checksummed or lowercase)"
                    },
                    "resource": {
                      "type": "string"
                    },
                    "effect": {
                      "type": "string",
                      "enum": [
                        "ALLOW",
                        "DENY"
                      ]
                    },
                    "created": {
                      "type": "boolean"
                    },
                    "removed": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden — requester does not have permission",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Forbidden — requester does not have permission",
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Internal server error",
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ]
}